+++ crud-vendas-demo/src/main/java/com/leo/crud/vendas/projections/SaleReportProjectionRecord.java
@@ -1,4 +1,10 @@
+//package com.leo.crud.vendas.projections;
+//
+//public record SaleReportProjectionRecord(String product, Long amount, Double unitValue, Double total) {
+//}
 package com.leo.crud.vendas.projections;
 
-public record SaleReportProjectionRecord(String product, Long amount, Double unitValue, Double total) {
+import java.math.BigDecimal;
+
+public record SaleReportProjectionRecord(String product, Long amount, BigDecimal unitValue, BigDecimal total) {
 }
